HTML_Database_SQLAlchemy with
Functional and stable project, some functions need to be slightly modified and add feedback when user modifies / creates items in database.

The html and css works fine for the moment.

I use a template with blocks in the html to make it much more agile when it comes to understanding the (html) code